+++ commands/issue.js
@@ -1,19 +1,7 @@
-const {sendGitHubEmbedReply} = require("../common.js");
-const request = require("request");
-module.exports = {
+const {gitHubSearchCommand} = require("../common.js");
+module.exports = gitHubSearchCommand({
     name: "issue",
     aliases: ["bug", "problem", "request"],
-    usage: "[search term]",
+    usage: "<search term>",
     description: "Search Minetest issues on GitHub",
-    execute: function(message, args) {
-        request({
-            url: "https://api.github.com/search/issues?q=is:issue+repo:minetest/minetest+" + encodeURI(args.join(" ")),
-            json: true,
-            headers: {
-                "User-Agent": "Minetest Bot"
-            }
-        }, function(err, res, pkg) {
-            sendGitHubEmbedReply(message, pkg.items[0]);
-        });
-    }
-}
+}, "issue");